varius fix

+++ Edi.Core/Gallery/Index/IndexGallery.cs
@@ -11,7 +11,7 @@ namespace Edi.Core.Gallery.Index
         public string Name { get; set; }
         public string Variant { get; set; }
         public int Duration { get; set; }
-        public int StartTime { get; set; }
+        public long StartTime { get; set; }
         public bool Loop { get; set; }
 
         public Dictionary<string, FileInfo> Assets { get; set; } = new Dictionary<string, FileInfo>();
